Move Button2 to main bundle
