+++ src/stories/ActionMenu/fixtures.stories.tsx
@@ -1,7 +1,18 @@
 import React from 'react'
 import {Meta} from '@storybook/react'
-import {ThemeProvider, BaseStyles, Box, Text, TextInput, StyledOcticon, FormGroup, ActionMenu, ActionList} from '../..'
-import {Button, IconButton} from '../../drafts'
+import {
+  ThemeProvider,
+  BaseStyles,
+  Box,
+  Text,
+  TextInput,
+  StyledOcticon,
+  FormGroup,
+  ActionMenu,
+  ActionList,
+  Button,
+  IconButton
+} from '../..'
 import {
   ServerIcon,
   PlusCircleIcon,
